BJS2-33477
refactoring code
