+++ src/main/java/school/faangSprint2/t11/DictionaryProcessor.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.function.BiConsumer;
 
 public class DictionaryProcessor {
-    private final Map<String, String> dictionary;
+    static Map<String, String> dictionary;
 
     public DictionaryProcessor() {
         this.dictionary = new HashMap<>();
@@ -13,18 +13,5 @@ public class DictionaryProcessor {
 
     public void processWord(String word, String translation, BiConsumer<String, String> consumer) {
         consumer.accept(word, translation);
-    }
-
-    public static void main(String[] args) {
-        DictionaryProcessor dictionaryProcessor = new DictionaryProcessor();
-
-        BiConsumer<String, String> addWordToDictionary = (word, translation) ->
-                dictionaryProcessor.dictionary.put(word, translation);
-
-        dictionaryProcessor.processWord("привет", "hello", addWordToDictionary);
-        dictionaryProcessor.processWord("мир", "world", addWordToDictionary);
-        dictionaryProcessor.processWord("программирование", "programming", addWordToDictionary);
-
-        System.out.println("Словарь: " + dictionaryProcessor.dictionary);
     }
 }
