+++ src/main/java/school/faangSprint2/t11/Main.java
@@ -1,0 +1,18 @@
+package school.faangSprint2.t11;
+
+import java.util.function.BiConsumer;
+
+public class Main {
+    public static void main(String[] args) {
+        DictionaryProcessor dictionaryProcessor = new DictionaryProcessor();
+
+        BiConsumer<String, String> addWordToDictionary = (word, translation) ->
+                DictionaryProcessor.dictionary.put(word, translation);
+
+        dictionaryProcessor.processWord("привет", "hello", addWordToDictionary);
+        dictionaryProcessor.processWord("мир", "world", addWordToDictionary);
+        dictionaryProcessor.processWord("программирование", "programming", addWordToDictionary);
+
+        System.out.println("Словарь: " + dictionaryProcessor.dictionary);
+    }
+}
